ready for test production TEST

+++ src/utils/product-create-err-handler.js
@@ -1,23 +1,21 @@
 const fs = require("fs");
 
 exports.handleCreateError = (req, res, err) => {
-  // Delete uploaded files if error occurs
   if (req.files) {
     req.files.forEach((file) => {
-      fs.unlink(file.path, (unlinkErr) => {
-        if (unlinkErr) {
-          console.error("Error deleting file:", unlinkErr);
-        }
-      });
+      if (file.path) {
+        fs.unlink(file.path, (unlinkErr) => {
+          if (unlinkErr) {
+            console.error("Error deleting file:", unlinkErr);
+          }
+        });
+      }
     });
   }
-  res.status(500).json({
-    error: ` ${err.message}`,
-  });
+  res.status(500).json({ error: ` ${err.message}` });
 };
 
 exports.handleUpdateError = (req, res, err) => {
-  // Delete uploaded files if error occurs
   if (req.file) {
     fs.unlink(req.file.path, (unlinkErr) => {
       if (unlinkErr) {
@@ -28,5 +26,15 @@ exports.handleUpdateError = (req, res, err) => {
 
   res.status(500).json({
     error: ` ${err.message}`,
+  });
+};
+
+exports.cleanupFiles = (filePaths) => {
+  filePaths.forEach((filePath) => {
+    fs.unlink(filePath, (unlinkErr) => {
+      if (unlinkErr) {
+        console.error("Error deleting file:", unlinkErr);
+      }
+    });
   });
 };
